+++ R/SAGE-marginal.R
@@ -23,14 +23,16 @@ MarginalSAGE = R6Class(
 	public = list(
 		#' @description
 		#' Creates a new instance of the MarginalSAGE class.
-		#' @param task,learner,measure,resampling,features,n_permutations,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
+		#' @param task,learner,measure,resampling,features,estimator,n_permutations,n_coalitions,batch_size,n_samples,early_stopping,se_threshold,min_permutations,check_interval Passed to [SAGE].
 		initialize = function(
 			task,
 			learner,
 			measure = NULL,
 			resampling = NULL,
 			features = NULL,
-			n_permutations = 10L,
+			estimator = c("permutation", "kernel"),
+			n_permutations = NULL,
+			n_coalitions = NULL,
 			batch_size = 5000L,
 			n_samples = 100L,
 			early_stopping = FALSE,
@@ -45,7 +47,9 @@ MarginalSAGE = R6Class(
 				measure = measure,
 				resampling = resampling,
 				features = features,
+				estimator = estimator,
 				n_permutations = n_permutations,
+				n_coalitions = n_coalitions,
 				batch_size = batch_size,
 				n_samples = n_samples,
 				early_stopping = early_stopping,